From importing mesh to normalization
Included the project database of the meshes
Read_data is to read all the data of the database, and store the information in an excel file

Preprocessing is to reduce the faces when it is too big (no mesh that was too small)

Feature_extraction is to make sure the longest part is on the x-axis, second longest on y-axis etc. Some meshes are not working out

Flipping is to check whether the mesh needs to be flipped. However, the final step of matrix multiplication is not finished.  Because  3x3 matrix times the flipping 4x4 won't work